create /tests for store backend tests
convert /tests into a python package

refactor requirements into folder

remove unused statesments

squashed commit

added flask min version to requirements

added sqlalchemy min version to requirements

Remove unused imports
