+++ .github/workflows/testbackend.yml
@@ -42,6 +42,7 @@ jobs:
     
     - name: Test with pytest
       run: |
+        PYTHONPATH="${PYTHONPATH}:/home/runner/work/Todolist/Todolist/backend/"
         export TESTING_DATABASE_URI=${{ secrets.TESTING_DATABASE_URI }}
         export SECRET_KEY=${{ secrets.SECRET_KEY }}
         pytest
